requirements.txt
chardet==3.0.4
Django==3.0.6
python-dotenv==0.13.0
requests==2.23.0
